made array of subsets

+++ SubsetRecord.java
@@ -1,12 +1,12 @@
 public class SubsetRecord {
     int numberOfBasicTerms; // n
-    int product; // product of selectivities
+    double product; // product of selectivities
     boolean noBranch;
     double bestCost;
     int leftChild;
     int rightChild;
 
-    public SubsetRecord(int n, int p) {
+    public SubsetRecord(int n, double p) {
         numberOfBasicTerms = n; // {f1, f2, f3, f4}
         product = p;
         noBranch = false;
